Refactored names according to etiquette
