--- src/main/java/inf112/skeleton/app/GUI/Pieces/Piece.java
@@ -1,18 +1,0 @@
-package inf112.skeleton.app.GUI.Pieces;
-
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.Batch;
-import com.badlogic.gdx.graphics.g2d.Sprite;
-import com.badlogic.gdx.scenes.scene2d.Actor;
-
-public class Piece extends Actor {
-
-    Sprite sprite;
-
-    @Override
-    public void draw(Batch batch, float parentAlpha) {
-        batch.draw(sprite, getX(), getY(), getWidth(), getHeight());
-
-    }
-}
