--- src/main/java/inf112/skeleton/app/Game/Enum/Direction.java
@@ -1,5 +1,0 @@
-package inf112.skeleton.app.Game.Enum;
-
-public enum Direction {
-    NORTH, SOUTH, EAST, WEST
-}
