--- src/main/java/inf112/skeleton/app/Game/Enum/Rotation.java
@@ -1,10 +1,0 @@
-package inf112.skeleton.app.Game.Enum;
-
-public enum Rotation {
-    /**
-     * L = left
-     * R = right
-     * U = U-turn
-     */
-    L, R, U
-}
