--- src/main/java/inf112/skeleton/app/Game/board/IActiveCell.java
@@ -1,11 +1,0 @@
-package inf112.skeleton.app.Game.board;
-
-public interface IActiveCell extends ICell{
-
-    /**
-     * Returns the action this cell does
-     * @return Action type of action
-     */
-    String getAction();
-
-}
