--- src/main/java/inf112/skeleton/app/Game/board/IBoard.java
@@ -1,16 +1,0 @@
-package inf112.skeleton.app.Game.board;
-
-public interface IBoard {
-
-    /**
-     * A method for generating the game board.
-     */
-    void generateBoard();
-
-    /**
-     * Prints out the board
-     * @param brett 2D array to be printed.
-     */
-    void displayBoard(char[][] brett);
-
-}
