+++ src/main/java/inf112/skeleton/app/gameLogic/board/IActiveCell.java
@@ -1,0 +1,11 @@
+package inf112.skeleton.app.gameLogic.board;
+
+public interface IActiveCell extends ICell{
+
+    /**
+     * Returns the action this cell does
+     * @return Action type of action
+     */
+    String getAction();
+
+}
