+++ src/main/java/inf112/skeleton/app/gameLogic/board/pieces/Laser.java
@@ -1,0 +1,42 @@
+package inf112.skeleton.app.gameLogic.board.pieces;
+
+import inf112.skeleton.app.gameLogic.enums.Direction;
+import inf112.skeleton.app.gameLogic.board.IPiece;
+
+public class Laser implements IPiece {
+
+    private Direction direction;
+    private LaserShooter laserShooter;
+
+    public Laser(Direction dir, LaserShooter laserShooter){
+        this.direction = dir;
+        this.laserShooter = laserShooter;
+    }
+
+    public LaserShooter getLaserShooter() {
+        return laserShooter;
+    }
+
+    public int getDamage(){
+        return laserShooter.getDamage();
+    }
+    @Override
+    public IPiece getType() {
+        return null;
+    }
+
+    @Override
+    public String getName() {
+        return "Laser";
+    }
+
+    @Override
+    public char getSymbol() {
+        return '-';
+    }
+
+    @Override
+    public Direction getRotation() {
+        return direction;
+    }
+}
