+++ src/main/java/inf112/skeleton/app/gameLogic/enums/Direction.java
@@ -1,0 +1,5 @@
+package inf112.skeleton.app.gameLogic.enums;
+
+public enum Direction {
+    NORTH, SOUTH, EAST, WEST
+}
