+++ src/main/java/inf112/skeleton/app/gameLogic/enums/Rotation.java
@@ -1,0 +1,10 @@
+package inf112.skeleton.app.gameLogic.enums;
+
+public enum Rotation {
+    /**
+     * L = left
+     * R = right
+     * U = U-turn
+     */
+    L, R, U
+}
